started matrix class
started added data creation commands

cleaned and improved netbuilder and network creation

just found out javafx does automatically render newlines...

started array parser for command line

started breadcrumbs of possible postgre-spring online server for record submission on exercises

+++ src/main/java/gk646/jnet/neuralnetwork/Neuron.java
@@ -4,16 +4,16 @@ import gk646.jnet.neuralnetwork.builder.NeuronInitState;
 
 public final class Neuron {
 
-    float bias;
+    double bias;
 
-    Neuron(float bias) {
+    Neuron(double bias) {
         this.bias = bias;
     }
 
     public static Neuron[] layer(int neuronCount, NeuronInitState neuronInitState) {
         Neuron[] temp = new Neuron[neuronCount];
         for (byte i = 0; i < neuronCount; i++) {
-            temp[i] = new Neuron(NetworkUtils.rng.nextFloat(neuronInitState.getOrigin(), neuronInitState.getBound()));
+            temp[i] = new Neuron(NetworkUtils.rng.nextDouble(neuronInitState.getOrigin(), neuronInitState.getBound()));
         }
         return temp;
     }
